fixed tests, refactored populator, updated docs
